+++ packages/credential-sdk/src/modules/tests/did-module.js
@@ -108,7 +108,7 @@ export default function generateDIDModuleTests(
       const service1 = new ServiceEndpoint('LinkedDomains', [
         'ServiceEndpoint#1',
       ]);
-      const service2 = new ServiceEndpoint('LinkedDomains', [
+      const service2 = new ServiceEndpoint('DIDCommMessaging', [
         'ServiceEndpoint#2',
       ]);
 
@@ -131,7 +131,7 @@ export default function generateDIDModuleTests(
       const service1 = new ServiceEndpoint('LinkedDomains', [
         'ServiceEndpoint#1',
       ]);
-      const service2 = new ServiceEndpoint('LinkedDomains', [
+      const service2 = new ServiceEndpoint('DIDCommMessaging', [
         'ServiceEndpoint#2',
       ]);