+++ examples/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @docknetwork/sdk-examples
 
+## 0.19.0
+
+### Minor Changes
+
+- Support DIDComm service endpoint
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/cheqd-blockchain-modules@2.0.0
+  - @docknetwork/credential-sdk@0.52.0
+  - @docknetwork/cheqd-blockchain-api@2.0.0
+
 ## 0.18.10
 
 ### Patch Changes