+++ packages/cheqd-blockchain-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docknetwork/cheqd-blockchain-api
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.52.0
+
 ## 1.0.0
 
 ### Minor Changes